Feat: 이력서 유형 추가
이력서 유형 추가 개발
